+++ src/components/map/BaseMap.tsx
@@ -13,7 +13,7 @@ import 'maplibre-gl/dist/maplibre-gl.css';
 
 import { DeckGLOverlay } from './DeckGLOverlay';
 import { EntityTooltip } from './EntityTooltip';
-import { WeatherTooltip, useWeatherPickerLayer } from './layers/WeatherOverlay';
+import { WeatherTooltip, useWeatherLayers } from './layers/WeatherOverlay';
 import { ThreatTooltip, useThreatHeatmapLayers } from './layers/ThreatHeatmapOverlay';
 import type { ThreatZoneData } from './layers/ThreatHeatmapOverlay';
 import { UtcClock } from '@/components/layout/UtcClock';
@@ -45,7 +45,6 @@ import { CompassControl } from './CompassControl';
 import { ProximityAlertOverlay } from './ProximityAlertOverlay';
 import { MapLegend } from './MapLegend';
 import { GeographicOverlay } from './layers/GeographicOverlay';
-import { PoliticalOverlay } from './layers/PoliticalOverlay';
 import { WeatherHeatmap } from './layers/WeatherHeatmap';
 
 /** Watches notificationStore.flyToTarget and animates the map. Renders null. */
@@ -88,7 +87,7 @@ export function BaseMap() {
   const setProximityPin = useFilterStore((s) => s.setProximityPin);
   const setSettingPin = useFilterStore((s) => s.setSettingPin);
   const entityLayers = useEntityLayers();
-  const weatherLayers = useWeatherPickerLayer();
+  const weatherLayers = useWeatherLayers();
   const threatLayers = useThreatHeatmapLayers();
   const isWeatherActive = useLayerStore((s) => s.activeLayers.has('weather'));
   const isThreatActive = useLayerStore((s) => s.activeLayers.has('threat'));
@@ -267,7 +266,6 @@ export function BaseMap() {
         />
         <WeatherHeatmap />
         <GeographicOverlay />
-        <PoliticalOverlay />
         <NavigationControl
           showZoom={true}
           showCompass={true}